add TypeScript parsing
