+++ package.json
@@ -12,6 +12,9 @@
     "lint": "eslint ./hunts/src/**"
   },
   "devDependencies": {
+    "@babel/preset-typescript": "^7.26.0",
+    "@types/react": "^19.0.4",
+    "@types/react-dom": "^19.0.2",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.25.3",
@@ -21,6 +24,7 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "prettier": "^2.2.1",
     "source-map-loader": "^4.0.1",
+    "typescript": "^5.7.3",
     "webpack-merge": "^5.7.0"
   },
   "dependencies": {
